checkpath.py: Set explict encoding in do_authors_exist.
It has  been discovered that  Python   versions  below 3.7  may incorrectly
detect  the encoding of text  files. If the system encoding   is not UTF-8,
then when building ovs with Python versions below 3.7, the ’24:checkpatch -
AUTHORS.rst existence’ test may  fail  with the  error ’UnicodeDecodeError:
'ascii' codec can't decode…’.

Reproduced in an environment  with the environment variable LC_CPTYPE  = C,
openvsitch 3.3.5, Python 3.6. Not relevant when using Python version 3.7 or
higher.  Starting  with this  version,   the  locale is converted  to UTF-8
regardless   of  the value of  environment   variables  for  Unix. Starting
from 3.15 UTF-8 mode is default for all platforms.

Signed-off-by: Ivan Burnin <iburnin@k2.cloud>